Выделит логику попапа фильма в отдельный презентер
